+++ src/Console/Command/AbstractConfigOptionCommand.php
@@ -11,7 +11,7 @@
  *
  * @package inetprocess/sugarcrm
  *
- * @license GNU General Public License v2.0
+ * @license Apache License 2.0
  *
  * @link http://www.inetprocess.com
  */
@@ -20,6 +20,7 @@ namespace SugarCli\Console\Command;
 
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Input\InputInterface;
+use Inet\SugarCRM\System as SugarSystem;
 
 abstract class AbstractConfigOptionCommand extends AbstractContainerAwareCommand
 {
@@ -90,5 +91,16 @@ abstract class AbstractConfigOptionCommand extends AbstractContainerAwareCommand
         }
 
         return $config->get($defaults[$name]);
+    }
+
+    /**
+     * Launch a simple quick repair and rebuild
+     */
+    protected function doQuickRepair($output)
+    {
+        // Can't call the command directly because of dependency injection
+        $sugarSystem = new SugarSystem($this->getService('sugarcrm.entrypoint'));
+        $sugarSystem->repair();
+        $output->writeln('<info>Repair Done.</info>');
     }
 }