+++ src/Console/Command/CodeSetupComposerCommand.php
@@ -11,19 +11,17 @@
  *
  * @package inetprocess/sugarcrm
  *
- * @license GNU General Public License v2.0
+ * @license Apache License 2.0
  *
  * @link http://www.inetprocess.com
  */
 
 namespace SugarCli\Console\Command;
 
-use Symfony\Component\Console\Input\ArrayInput;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Finder\Finder;
-use Inet\SugarCRM\System as SugarSystem;
 
 class CodeSetupComposerCommand extends AbstractConfigOptionCommand
 {
@@ -89,20 +87,20 @@ class CodeSetupComposerCommand extends AbstractConfigOptionCommand
         } elseif (empty($composerUtil) && !empty($composerJson)) {
             $msg = "Looks like you have a composer.json but you don't have the Util.";
         } elseif (empty($composerUtil) && empty($composerJson)) {
-            $msg = "Looks like you have neither a composer.json nor the Util.";
+            $msg = 'Looks like you have neither a composer.json nor the Util.';
         } elseif ($input->getOption('reinstall') === true) {
-            $msg = "Everything is installed but will reinstall";
+            $msg = 'Everything is installed but will reinstall';
         } else {
-            $output->writeln("<info>Everything seems fine ! Used --reinstall to reinstall</info>");
+            $output->writeln('<info>Everything seems fine ! Used --reinstall to reinstall</info>');
+
             return;
         }
 
         $output->writeln("<comment>$msg</comment>");
 
         if ($input->getOption('reinstall') === false) {
-            $output->writeln("<comment>Will install it (require --do to have an effect)</comment>");
+            $output->writeln('<comment>Will install it (require --do to have an effect)</comment>');
         }
-
 
         // create the composer Util
         if ($input->getOption('reinstall') === true || (empty($composerUtil) && $input->getOption('do') === true)) {
@@ -114,73 +112,44 @@ class CodeSetupComposerCommand extends AbstractConfigOptionCommand
             $output->writeln(PHP_EOL . $this->createComposerJson($sugarPath));
         }
 
-        $output->writeln("<info>Job done !</info>");
+        $output->writeln('<info>Job done !</info>');
         if ($input->getOption('no-quickrepair') === false) {
-            $output->writeln("Launching a quick repair and rebuild</info>");
+            $output->writeln('Launching a quick repair and rebuild</info>');
             $this->doQuickRepair($output);
         }
     }
 
     /**
      * Create the composer.php util
-     * @return    string
+     *
+     * @return string
      */
     protected function createComposerUtil($utilsPath)
     {
-        file_put_contents($utilsPath . '/composer.php', "<?php
-require_once(__DIR__ . '/../../../vendor/autoload.php');
-");
+        copy(__DIR__ . '/../../../res/code_templates/composer.php', "{$utilsPath}/composer.php");
 
         return " --> Util installed in $utilsPath/composer.php";
     }
 
     /**
      * Create the composer.json file
-     * @return    string
+     *
+     * @return string
      */
     protected function createComposerJson($sugarPath)
     {
         $phpVersion = explode('.', PHP_VERSION);
         $phpVersion = "{$phpVersion[0]}.{$phpVersion[1]}";
-        $data = '{
-    "require-dev": {
-        "inetprocess/libsugarcrm": "^1-beta"
-    },
-    "autoload": {
-        "psr-0": {
-            "": "include/lib/"
-        }
-    },
-    "autoload-dev": {
-        "psr-4": {
-            "Inet\\\SugarCRM\\\Tests\\\": "vendor/inetprocess/libsugarcrm/tests/"
-        }
-    },
-    "config": {
-         "platform": {
-            "php": "' . $phpVersion . '"
-         }
-    }
-}
-';
-        file_put_contents($sugarPath . '/custom/composer.json', $data);
+
+        $jsonContent = file_get_contents(__DIR__ . '/../../../res/code_templates/composer.json');
+        // replace the vars
+        $jsonContent = str_replace('[[PHP_VERSION]]', $phpVersion, $jsonContent);
+        file_put_contents($sugarPath . '/custom/composer.json', $jsonContent);
         $msg = " --> composer.json installed in custom/composer.json for PHP $phpVersion" . PHP_EOL;
-        $msg.= "     You can now use composer require package/name to install packages" . PHP_EOL;
-        $msg.= "     Put all your classes (with namespaces) in custom/include/lib" . PHP_EOL;
-        $msg.= "     Finally, do a composer install to install all the components" . PHP_EOL;
+        $msg.= '     You can now use composer require package/name to install packages' . PHP_EOL;
+        $msg.= '     Put all your classes (with namespaces) in custom/include/lib' . PHP_EOL;
+        $msg.= '     Finally, do a composer install to install all the components' . PHP_EOL;
 
         return $msg;
-    }
-
-    /**
-     * Launch a simple quick repair and rebuild
-     */
-    protected function doQuickRepair($output)
-    {
-        // Can't call the command directly because of dependency injection
-        $sugarSystem = new SugarSystem($this->getService('sugarcrm.entrypoint'));
-        $sugarSystem->repair();
-        $output->writeln('<info>Repair Done.</info>');
-
     }
 }